rename VirtualMachine to AzureVirtualMachine
