+++ cartography/intel/azure/subscription.py
@@ -6,7 +6,7 @@ import neo4j
 from azure.core.exceptions import HttpResponseError
 from azure.mgmt.resource import SubscriptionClient
 
-from .credentials import Credentials
+from .util.credentials import Credentials
 from cartography.util import run_cleanup_job
 from cartography.util import timeit
